requirements.txt
fastapi
mistralai
uvicorn
pydantic
python-dotenv